fixed bug where self_sign_log_archives report action was causing a 500 error.
report results were not being filtered as the post request data array was getting blackholed by the cakephp security component.

turned off the security component for that post request.

added logic to check that a filter of some kind is set before generating a report and that report results do not exceed 5000