Add controller method to handle duplication logic with the first layer of duplication
The controller will handle all of the duplication logic (possibly with
the help of helper methods). So far the first layer of logic is working,
that is, the program itself being duplicated. The transactionIds array
will hold all of the id's nested under an associative array bearing the
model type itself as keys (e.g Program, ProgramStep, ProgramDocument)

[otf: 37 wl: 13 minutes]