updated 'old' design architecture logic
Added /webroot/img/default to master repo so that themes could deploy without crashing. Upon doing so, when using the old architecture - it would attempt to mv /design/img/default to /webroot/img/default - and now /webroot/img/default exists, so it fails. Added logic so that if it does exist, it removes the existing first.